requirements.txt
requests
python-dotenv
beautifulsoup4
webdriver-manager
selenium
loguru